# Parsers returns multiple items & Reuse Items buffer instead of streams 