--- a/application/apps/indexer/addons/dlt-tools/src/lib.rs
+++ b/application/apps/indexer/addons/dlt-tools/src/lib.rs
@@ -16,7 +16,6 @@ extern crate indexer_base;
 extern crate log;
 
 use dlt_core::filtering::DltFilterConfig;
-use futures::pin_mut;
 use parsers::{dlt::DltParser, Attachment, MessageStreamItem, ParseYield};
 use sources::{binary::raw::BinaryByteSource, producer::MessageProducer};
 use std::{
@@ -44,8 +43,6 @@ pub async fn scan_dlt_ft(
                 with_storage_header,
             );
             let mut producer = MessageProducer::new(parser, source, None);
-            let stream = producer.as_stream();
-            pin_mut!(stream);
 
             let mut canceled = false;
 
@@ -57,14 +54,14 @@ pub async fn scan_dlt_ft(
                         canceled = true;
                         break;
                     }
-                    items = tokio_stream::StreamExt::next(&mut stream) => {
+                    items = producer.read_next_segment() => {
                         match items {
                             Some(items) => {
                                 for (_, item) in items {
                                     if let MessageStreamItem::Item(ParseYield::MessageAndAttachment((_msg, attachment))) = item {
-                                        attachments.push(attachment);
+                                        attachments.push(attachment.to_owned());
                                     } else if let MessageStreamItem::Item(ParseYield::Attachment(attachment)) = item {
-                                        attachments.push(attachment);
+                                        attachments.push(attachment.to_owned());
                                     }
                                 }
                             }