--- a/application/apps/indexer/indexer_cli/src/main.rs
+++ b/application/apps/indexer/indexer_cli/src/main.rs
@@ -31,7 +31,6 @@ use dlt_core::{
 };
 use dlt_tools::{extract_dlt_ft, scan_dlt_ft};
 use env_logger::Env;
-use futures::{pin_mut, stream::StreamExt};
 use indexer_base::config::*;
 use indicatif::{ProgressBar, ProgressStyle};
 use parsers::{
@@ -806,10 +805,10 @@ pub async fn main() -> Result<()> {
             let dlt_parser = DltParser::new(None, None, None, None, true);
             let reader = BufReader::new(&in_file);
             let source = BinaryByteSource::new(reader);
-            let mut dlt_msg_producer = MessageProducer::new(dlt_parser, source, None);
+            let dlt_msg_producer = MessageProducer::new(dlt_parser, source, None);
             let cancel = CancellationToken::new();
             export_raw(
-                Box::pin(dlt_msg_producer.as_stream()),
+                dlt_msg_producer,
                 &out_path,
                 &sections,
                 false,
@@ -1372,8 +1371,12 @@ async fn count_dlt_messages(input: &Path) -> Result<u64, DltParseError> {
         let source = BinaryByteSource::new(second_reader);
 
         let mut dlt_msg_producer = MessageProducer::new(dlt_parser, source, None);
-        let msg_stream = dlt_msg_producer.as_stream();
-        Ok(msg_stream.count().await as u64)
+        let mut msgs_count = 0;
+        while let Some(items) = dlt_msg_producer.read_next_segment().await {
+            msgs_count += items.len();
+        }
+
+        Ok(msgs_count as u64)
     } else {
         Err(DltParseError::Unrecoverable(format!(
             "Couldn't find dlt file: {input:?}"
@@ -1389,13 +1392,11 @@ async fn detect_messages_type(input: &Path) -> Result<bool, DltParseError> {
             let source = BinaryByteSource::new(buf_reader);
             let dlt_parser = DltRangeParser::new();
             let mut dlt_msg_producer = MessageProducer::new(dlt_parser, source, None);
-            let msg_stream = dlt_msg_producer.as_stream();
-            pin_mut!(msg_stream);
             let mut item_count = 0usize;
             let mut attachment_count = 0usize;
             let mut err_count = 0usize;
             let mut consumed = 0usize;
-            'outer: while let Some(items) = msg_stream.next().await {
+            'outer: while let Some(items) = dlt_msg_producer.read_next_segment().await {
                 for item in items {
                     match item {
                         (_, MessageStreamItem::Item(ParseYield::Message(item))) => {
@@ -1436,18 +1437,16 @@ async fn detect_messages_type(input: &Path) -> Result<bool, DltParseError> {
             match PcapngByteSource::new(fs::File::open(input)?) {
                 Ok(source) => {
                     let mut some_msg_producer = MessageProducer::new(some_parser, source, None);
-                    let msg_stream = some_msg_producer.as_stream();
-                    pin_mut!(msg_stream);
                     let mut item_count = 0usize;
                     let mut err_count = 0usize;
                     let mut consumed = 0usize;
                     let mut skipped_count = 0usize;
-                    'outer: while let Some(items) = msg_stream.next().await {
+                    'outer: while let Some(items) = some_msg_producer.read_next_segment().await {
                         for item in items {
                             match item {
                                 (used, MessageStreamItem::Item(_)) => {
                                     item_count += 1;
-                                    consumed += used;
+                                    consumed += *used;
                                 }
                                 (_, MessageStreamItem::Skipped) => skipped_count += 1,
                                 (_, MessageStreamItem::Incomplete) => err_count += 1,
@@ -1475,14 +1474,12 @@ async fn detect_messages_type(input: &Path) -> Result<bool, DltParseError> {
             match PcapngByteSource::new(fs::File::open(input)?) {
                 Ok(source) => {
                     let mut dlt_msg_producer = MessageProducer::new(dlt_parser, source, None);
-                    let msg_stream = dlt_msg_producer.as_stream();
-                    pin_mut!(msg_stream);
                     let mut item_count = 0usize;
                     let mut attachment_count = 0usize;
                     let mut err_count = 0usize;
                     let mut consumed = 0usize;
                     let mut skipped_count = 0usize;
-                    'outer: while let Some(items) = msg_stream.next().await {
+                    'outer: while let Some(items) = dlt_msg_producer.read_next_segment().await {
                         for item in items {
                             match item {
                                 (_, MessageStreamItem::Item(ParseYield::Message(item))) => {
@@ -1527,15 +1524,13 @@ async fn detect_messages_type(input: &Path) -> Result<bool, DltParseError> {
             let buf_reader = BufReader::new(fs::File::open(input)?);
             let source = BinaryByteSource::new(buf_reader);
             let mut txt_msg_producer = MessageProducer::new(txt_parser, source, None);
-            let msg_stream = txt_msg_producer.as_stream();
-            pin_mut!(msg_stream);
             let mut item_count = 0usize;
             let mut err_count = 0usize;
             let mut skipped_count = 0usize;
             let mut consumed = 0usize;
             let mut attachment_count = 0usize;
             use std::io::Cursor;
-            'outer: while let Some(items) = msg_stream.next().await {
+            'outer: while let Some(items) = txt_msg_producer.read_next_segment().await {
                 for item in items {
                     match item {
                         (_rest, MessageStreamItem::Item(ParseYield::Message(item))) => {