--- a/application/apps/indexer/parsers/src/lib.rs
+++ b/application/apps/indexer/parsers/src/lib.rs
@@ -3,7 +3,7 @@ pub mod dlt;
 pub mod someip;
 pub mod text;
 use serde::Serialize;
-use std::{fmt::Display, io::Write};
+use std::{fmt::Display, io::Write, iter};
 use thiserror::Error;
 
 extern crate log;
@@ -98,4 +98,57 @@ pub enum MessageStreamItem<T: LogMessage> {
     Incomplete,
     Empty,
     Done,
+}
+
+/// Continuously applies the [`parse_fn`] function to the given [`input`] bytes,
+/// extracting all possible items until no more can be parsed.
+///
+/// This function serves as a helper for parsing methods that return only the first
+/// successfully parsed item. It repeatedly invokes [`parse_fn`] as long as there
+/// are enough bytes remaining in [`input`] to parse another message.
+///
+/// # Behavior
+///
+/// - If the first invocation of [`parse_fn`] fails, the function returns an error immediately.
+/// - If the first invocation succeeds, parsing continues until either:
+///   - There are not enough remaining bytes to parse another message.
+///   - [`parse_fn`] returns an error, which will be ignored after the first success.
+///
+/// # Arguments
+///
+/// * `input`: A slice of bytes to be parsed.
+/// * `timestamp`: An optional timestamp associated with the message.
+/// * `min_bytes_count`: The minimum number of bytes likely required to parse a message.
+/// * `parse_fn`: A function that attempts to parse a message from a byte slice.
+///   - It takes a byte slice and an optional timestamp.
+///   - It returns a result containing:
+///     - The number of bytes consumed.
+///     - An optional parsed item [`ParseYield<T>`].
+///   - If parsing fails, it returns an [`Error`].
+fn parse_all<F, T>(
+    input: &[u8],
+    timestamp: Option<u64>,
+    min_bytes_count: usize,
+    mut parse_fn: F,
+) -> Result<impl Iterator<Item = (usize, Option<ParseYield<T>>)> + use<'_, F, T>, Error>
+where
+    F: FnMut(&[u8], Option<u64>) -> Result<(usize, Option<ParseYield<T>>), Error>,
+{
+    let mut slice = input;
+
+    // return early if function errors on first parse call.
+    let first_res = parse_fn(slice, timestamp)?;
+
+    // Otherwise keep parsing and stop on first error, returning the parsed items at the end.
+    let iter = iter::successors(Some(first_res), move |(consumed, _res)| {
+        slice = &slice[*consumed..];
+
+        if slice.len() < min_bytes_count {
+            return None;
+        }
+
+        parse_fn(slice, timestamp).ok()
+    });
+
+    Ok(iter)
 }