--- a/application/apps/indexer/processor/src/export/mod.rs
+++ b/application/apps/indexer/processor/src/export/mod.rs
@@ -3,9 +3,9 @@ use std::{
     path::Path,
 };
 
-use futures::StreamExt;
 use indexer_base::config::IndexSection;
-use parsers::{LogMessage, MessageStreamItem, ParseYield};
+use parsers::{LogMessage, MessageStreamItem, ParseYield, Parser};
+use sources::{producer::MessageProducer, ByteSource};
 use thiserror::Error;
 use tokio_util::sync::CancellationToken;
 
@@ -39,8 +39,8 @@ pub enum ExportError {
 ///
 /// # Errors
 /// In case of cancellation will return ExportError::Cancelled
-pub async fn export_raw<S, T>(
-    mut s: S,
+pub async fn export_raw<T, P, D>(
+    mut producer: MessageProducer<T, P, D>,
     destination_path: &Path,
     sections: &Vec<IndexSection>,
     read_to_end: bool,
@@ -49,7 +49,8 @@ pub async fn export_raw<S, T>(
 ) -> Result<usize, ExportError>
 where
     T: LogMessage + Sized,
-    S: futures::Stream<Item = Box<[(usize, MessageStreamItem<T>)]>> + Unpin,
+    P: Parser<T>,
+    D: ByteSource,
 {
     trace!("export_raw, sections: {sections:?}");
     if !sections_valid(sections) {
@@ -70,7 +71,7 @@ where
     if sections.is_empty() {
         debug!("no sections configured");
         // export everything
-        'outer: while let Some(items) = s.next().await {
+        'outer: while let Some(items) = producer.read_next_segment().await {
             if cancel.is_cancelled() {
                 return Err(ExportError::Cancelled);
             }
@@ -99,7 +100,7 @@ where
         return Ok(exported);
     }
 
-    'outer: while let Some(items) = s.next().await {
+    'outer: while let Some(items) = producer.read_next_segment().await {
         if cancel.is_cancelled() {
             return Err(ExportError::Cancelled);
         }
@@ -150,7 +151,7 @@ where
         }
     }
     if read_to_end {
-        'outer: while let Some(items) = s.next().await {
+        'outer: while let Some(items) = producer.read_next_segment().await {
             if cancel.is_cancelled() {
                 return Err(ExportError::Cancelled);
             }