This PR introduces the following changes:

### Producer:
* Parsed items will be saved in an internal buffer reusing the memory to avoid costly memory allocation calls especially with parsers returning multiple items at one.
* Change producer API removing the stream interface and making `load_next_segment()` function returns a mutable reference of the newly loaded items, giving the caller the freedom to consume the items with `drain(..)` or iterate over their references.
* Unused field `index` is removed.
* Cancel safety if `load_next_segment()` is described in the comments in code and tested manually.
* Unit tests has been adjusted for the new producer API, and they all pass without any change in their core functionality. 

### Parsers:
* Make all parsers try to parse all possible items from the provided bytes instead of returning the first item only. If the first call of parse fails, then the error will be returned, however any errors after having a successful parse call will be ignored in the current call because parse will be called again with more data. If the error persists on the next call, then it will be returned directly since it will occur on the first parse call.
* Unit tests for all parsers adjusted to ensure the count of the parsed items on each call + New tests has been added.

**Note**: The checks on TCP & UDP byte sources introduced in #2230 aren't needed anymore for the built-in parser. However, I kept them to support plugin parser that can parse one item at a time only.

### Fixes:
* Mock parser for benchmarks avoids allocating memory by using `iter::repeate_n()` instead.
* Chipmunk CLI: Add warning for external dependencies.

----

## Performance Changes:
The changes led to slightly increasing in performance in spite of the parser returning multiple values. This improvement is thanks to reusing the memory with the internal buffer in producer loop. Here is a comparison between master and current PR.
```md
# Chipmunk App: Parsing DLT file 500MB:

### Current PR:
File Read Took: 12058 ms
File Read Took: 11972 ms
File Read Took: 12103 ms

### Master: 
File Read Took: 12648 ms
File Read Took: 12639 ms
File Read Took: 12680 ms

--------------------------------------------------

# Chipmunk App: Parsing DLT With attachments 150MB:

### Current PR:
File Read Took: 3509 ms
File Read Took: 3461 ms
File Read Took: 3538 ms

### Master: 
File Read Took: 3518 ms
File Read Took: 3485 ms
File Read Took: 3507 ms

--------------------------------------------------

# Rust Core Benchmarks: Mock parser returning 10 values on each call:

### Current PR:
time:   [5.9618 ms 5.9758 ms 5.9899 ms]
time:   [5.9676 ms 5.9816 ms 5.9954 ms]

### Master: 
time:   [8.3135 ms 8.3412 ms 8.3692 ms]
time:   [8.2700 ms 8.2920 ms 8.3141 ms]
```